+++ common/cosmwasm-smart-contracts/coconut-bandwidth-contract/src/deposit.rs
@@ -4,7 +4,7 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 pub struct DepositData {
     deposit_info: String,
     identity_key: String,