+++ explorer-api/src/http/mod.rs
@@ -92,9 +92,9 @@ fn get_servers() -> Vec<rocket_okapi::okapi::openapi3::Server> {
     if std::env::var_os("CARGO").is_some() {
         return vec![];
     }
-    return vec![rocket_okapi::okapi::openapi3::Server {
+    vec![rocket_okapi::okapi::openapi3::Server {
         url: std::env::var("OPEN_API_BASE").unwrap_or_else(|_| "/api/v1/".to_owned()),
         description: Some("API".to_owned()),
         ..Default::default()
-    }];
+    }]
 }